chore: Atualização dos nomes das pastas seguindo o padrão de features
