+++ src/Baltaio.Location.Api/Application/Addresses/Commons/IAddressRepository.cs
@@ -1,0 +1,8 @@
+﻿using Baltaio.Location.Api.Domain.Addresses;
+
+namespace Baltaio.Location.Api.Application.Addresses.Commons;
+
+public interface IAddressRepository
+{
+    public Task SaveAsync(Address address);
+}
