+++ src/Baltaio.Location.Api/Application/Addresses/Commons/ICityRepository.cs
@@ -1,0 +1,8 @@
+﻿using Baltaio.Location.Api.Domain.Addresses;
+
+namespace Baltaio.Location.Api.Application.Addresses.Commons;
+
+public interface ICityRepository
+{
+    public Task<City?> GetAsync(int ibgeCode);
+}
