+++ src/Baltaio.Location.Api/Application/Addresses/CreateAddress/CreateAddressAppService.cs
@@ -1,7 +1,7 @@
-﻿using Baltaio.Location.Api.Application.Commons;
-using Baltaio.Location.Api.Domain.Locations;
+﻿using Baltaio.Location.Api.Application.Addresses.Commons;
+using Baltaio.Location.Api.Domain.Addresses;
 
-namespace Baltaio.Location.Api.Application.CreateAddress;
+namespace Baltaio.Location.Api.Application.Addresses.CreateAddress;
 
 public class CreateAddressAppService
 {
