+++ src/Baltaio.Location.Api/Application/Addresses/CreateAddress/CreateAddressInput.cs
@@ -1,0 +1,3 @@
+﻿namespace Baltaio.Location.Api.Application.Addresses.CreateAddress;
+
+public record CreateAddressInput(int IbgeCode);
