+++ src/Baltaio.Location.Api/Application/Addresses/CreateAddress/CreateAddressOutput.cs
@@ -1,0 +1,3 @@
+﻿namespace Baltaio.Location.Api.Application.Addresses.CreateAddress;
+
+public record CreateAddressOutput(Guid AddressCode, string Message);
