--- src/Baltaio.Location.Api/Application/Commons/IAddressRepository.cs
@@ -1,8 +1,0 @@
-﻿using Baltaio.Location.Api.Domain.Locations;
-
-namespace Baltaio.Location.Api.Application.Commons;
-
-public interface IAddressRepository
-{
-    public Task SaveAsync(Address address);
-}
