--- src/Baltaio.Location.Api/Application/Commons/ICityRepository.cs
@@ -1,8 +1,0 @@
-﻿using Baltaio.Location.Api.Domain.Locations;
-
-namespace Baltaio.Location.Api.Application.Commons;
-
-public interface ICityRepository
-{
-    public Task<City?> GetAsync(int ibgeCode);
-}
