--- src/Baltaio.Location.Api/Application/CreateAddress/CreateAddressInput.cs
@@ -1,3 +1,0 @@
-﻿namespace Baltaio.Location.Api.Application.CreateAddress;
-
-public record CreateAddressInput(int IbgeCode);
