--- src/Baltaio.Location.Api/Application/CreateAddress/CreateAddressOutput.cs
@@ -1,3 +1,0 @@
-﻿namespace Baltaio.Location.Api.Application.CreateAddress;
-
-public record CreateAddressOutput(Guid AddressCode, string Message);
