+++ src/Baltaio.Location.Api/Controllers/Addresses/AddressesController.cs
@@ -1,5 +1,5 @@
-﻿using Baltaio.Location.Api.Application.Commons;
-using Baltaio.Location.Api.Application.CreateAddress;
+﻿using Baltaio.Location.Api.Application.Addresses.Commons;
+using Baltaio.Location.Api.Application.Addresses.CreateAddress;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Baltaio.Location.Api.Controllers.Addresses;
