+++ src/Baltaio.Location.Api/Domain/Addresses/City.cs
@@ -1,0 +1,5 @@
+﻿namespace Baltaio.Location.Api.Domain.Addresses;
+
+public sealed class City
+{
+}
