--- src/Baltaio.Location.Api/Domain/Locations/City.cs
@@ -1,5 +1,0 @@
-﻿namespace Baltaio.Location.Api.Domain.Locations;
-
-public sealed class City
-{
-}
