+++ tests/Baltaio.Location.Api.Tests/Domain/AddressTests.cs
@@ -1,4 +1,4 @@
-﻿using Baltaio.Location.Api.Domain.Locations;
+﻿using Baltaio.Location.Api.Domain.Addresses;
 using FluentAssertions;
 
 namespace Baltaio.Location.Api.Tests.Domain;
